+++ tlb/src/main/java/org/ton/java/tlb/BlockExtra.java
@@ -55,7 +55,7 @@ public class BlockExtra {
         .storeRef(outMsgDesc.toCell())
         .storeRef(
             shardAccountBlocks.serialize(
-                k -> CellBuilder.beginCell().storeUint((Long) k, 256).endCell().getBits(),
+                k -> CellBuilder.beginCell().storeUint((BigInteger) k, 256).endCell().getBits(),
                 v -> CellBuilder.beginCell().storeCell(((AccountBlock) v).toCell()),
                 e -> CellBuilder.beginCell().storeCell(((CurrencyCollection) e).toCell()),
                 (fk, fv) -> CellBuilder.beginCell().storeUint(((Long) fk) + ((Long) fv), 32)))
