+++ tlb/src/main/java/org/ton/java/tlb/ConfigParams45.java
@@ -1,5 +1,6 @@
 package org.ton.java.tlb;
 
+import java.math.BigInteger;
 import lombok.Builder;
 import lombok.Data;
 import org.ton.java.cell.Cell;
@@ -20,7 +21,7 @@ public class ConfigParams45 {
         .storeUint(0xc0, 8)
         .storeDict(
             precompiledContractsList.serialize(
-                k -> CellBuilder.beginCell().storeUint((Long) k, 256).endCell().getBits(),
+                k -> CellBuilder.beginCell().storeUint((BigInteger) k, 256).endCell().getBits(),
                 v -> CellBuilder.beginCell().storeCell(((PrecompiledSmc) v).toCell()).endCell()))
         .storeUint(suspendedUntil, 32)
         .endCell();
