+++ tlb/src/main/java/org/ton/java/tlb/JettonBridgeParamsV2.java
@@ -1,5 +1,6 @@
 package org.ton.java.tlb;
 
+import java.math.BigInteger;
 import lombok.Builder;
 import lombok.Data;
 import org.ton.java.address.Address;
@@ -7,8 +8,6 @@ import org.ton.java.cell.Cell;
 import org.ton.java.cell.CellBuilder;
 import org.ton.java.cell.CellSlice;
 import org.ton.java.cell.TonHashMapE;
-
-import java.math.BigInteger;
 
 @Builder
 @Data
@@ -29,7 +28,7 @@ public class JettonBridgeParamsV2 implements JettonBridgeParams {
         .storeAddress(oracleAddress)
         .storeDict(
             oracles.serialize(
-                k -> CellBuilder.beginCell().storeUint((Long) k, 256).endCell().getBits(),
+                k -> CellBuilder.beginCell().storeUint((BigInteger) k, 256).endCell().getBits(),
                 v -> CellBuilder.beginCell().storeUint((BigInteger) v, 256).endCell()))
         .storeUint(stateFlags, 8)
         .storeRef(prices.toCell())
