+++ tlb/src/main/java/org/ton/java/tlb/McStateExtraInfo.java
@@ -44,7 +44,7 @@ public class McStateExtraInfo {
         .storeCell(validatorInfo.toCell())
         .storeDict(
             prevBlocks.serialize(
-                k -> CellBuilder.beginCell().storeUint((Long) k, 32).endCell().getBits(),
+                k -> CellBuilder.beginCell().storeUint((BigInteger) k, 32).endCell().getBits(),
                 v -> CellBuilder.beginCell().storeCell(((KeyExtBlkRef) v).toCell()),
                 e -> CellBuilder.beginCell().storeCell(((KeyMaxLt) e).toCell()),
                 (fk, fv) -> CellBuilder.beginCell().storeUint(0, 1)))
