+++ tlb/src/main/java/org/ton/java/tlb/ShardAccounts.java
@@ -1,5 +1,6 @@
 package org.ton.java.tlb;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -27,7 +28,7 @@ public class ShardAccounts {
     return CellBuilder.beginCell()
         .storeCell(
             shardAccounts.serialize(
-                k -> CellBuilder.beginCell().storeUint((Long) k, 256).endCell().getBits(),
+                k -> CellBuilder.beginCell().storeUint((BigInteger) k, 256).endCell().getBits(),
                 v -> CellBuilder.beginCell().storeCell(((ShardAccount) v).toCell()),
                 e -> CellBuilder.beginCell().storeCell(((DepthBalanceInfo) e).toCell()),
                 (fk, fv) -> CellBuilder.beginCell().storeUint(0, 1)))
